upload game page update (using components)
