+++ src/App.js
@@ -6,10 +6,9 @@ import 'primeflex/primeflex.css'
 
 import TopBar from './components/layout/TopBar'
 import PlayedGames from './components/pages/PlayedGamesList'
-// import UploadGame from './components/pages/UploadGame'
+import UploadGame from './components/pages/UploadGame'
 import GameDetails from './components/pages/GameDetails'
 import GameStats from './components/pages/GameStats'
-// import FavoriteGames from './components/pages/FavoriteGames'
 import Home from './components/pages/Home'
 import Dashboard from './components/pages/Dashboard'
 import Login from './components/pages/Login'
@@ -42,6 +41,7 @@ function App() {
                   <Route path={routes.playedgames} element={<PlayedGames />} />
                   <Route path={routes.stats} element={<GameStats />} />
                   <Route path={routes.gamedetails} element={<GameDetails />} />
+                  <Route path={routes.uploadgame} element={<UploadGame />} />
                 </Routes>
               </PlayedGamesProvider>
             </MessagesProvider>
